+++ run_integ_tests.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -e
+
+# We need to change to the directory to pick up the cargo config
+(cd crates/integ && cargo nextest run)